Remove unnecessary specs
The test "display administrator id on public views" is not correct. The valuation comments
are not display never on public views. This test pass because there is an error at the
moment to render the comments. If we reload this admin page we can see that the
description is render instead of administrator_id.

The test "public users not see admin description"  is already being checked in the
'system/comments_specs'. However, we are going to add a new expectation to
make sure that the admin description does not appear on the public pages.